セキュリティ対策 (フォームから送信されたuseIdとサーバー側でDBに登録しているidが一致するか確認)

+++ app/action/ActionAppIntroduction.ts
@@ -3,9 +3,10 @@
 import { redirect } from "next/navigation";
 import prisma from "../lib/prisma";
 import { z } from "zod";
-import { fileSaveStorage } from "../lib/fileSaveStorage";
+import { fileSaveStorage } from "../lib/FileSaveStorage";
 import { validateMimeTypeAndExtension } from "../lib/validateMimeTypeAndExtension";
 import { revalidatePath } from "next/cache";
+import { getSessionUserId } from "../lib/sessionUserService";
 
 type FormState = {
   message?: string | null;
@@ -52,8 +53,16 @@ export const addAppIntroduction = async (
   const solution = formData.get("solution") as string;
   const image = formData.get("imageFile") as File;
   const imageALT = formData.get("imageALT") as string;
-
   const userId = formData.get("userId") as string;
+  const sessionUserId = await getSessionUserId();
+
+  if (!sessionUserId) {
+    throw new Error("セッションを取得できませんでした。");
+  }
+
+  if (sessionUserId !== userId) {
+    throw new Error("セッションとフォームで送信されたIDが一致しません。");
+  }
 
   const canArray = [];
   let canIndex = 0;
@@ -169,9 +178,36 @@ export const updateAppIntroduction = async (
   const solution = formData.get("solution") as string;
   const image = formData.get("imageFile") as File;
   const imageALT = formData.get("imageALT") as string;
-
   const appId = formData.get("appId") as string;
   const userId = formData.get("userId") as string;
+  const sessionUserId = await getSessionUserId();
+
+  if (!sessionUserId) {
+    throw new Error("セッションを取得できませんでした。");
+  }
+
+  if (sessionUserId !== userId) {
+    throw new Error("セッションとフォームで送信されたIDが一致しません。");
+  }
+
+  const appIntroduction = await prisma.appIntroduction.findUnique({
+    where: {
+      id: appId,
+    },
+    include: {
+      user: true,
+    },
+  });
+
+  if (!appIntroduction) {
+    throw new Error("登録しているアプリが見つかりません。");
+  }
+
+  if (appIntroduction.userId !== userId) {
+    throw new Error(
+      "登録しているユーザーのIDとフォームから送信されたユーザーのIDが一致しません。"
+    );
+  }
 
   const canArray = [];
   let canIndex = 0;
@@ -276,12 +312,41 @@ export const updateAppIntroduction = async (
 };
 
 export const deleteAppIntroduction = async (formData: FormData) => {
-  const id = formData.get("appId") as string;
+  const appId = formData.get("appId") as string;
+  const userId = formData.get("userId") as string;
+  const sessionUserId = await getSessionUserId();
+
+  if (!sessionUserId) {
+    throw new Error("セッションを取得できませんでした。");
+  }
+
+  if (sessionUserId !== userId) {
+    throw new Error("セッションとフォームで送信されたIDが一致しません。");
+  }
+
+  const appIntroduction = await prisma.appIntroduction.findUnique({
+    where: {
+      id: appId,
+    },
+    include: {
+      user: true,
+    },
+  });
+
+  if (!appIntroduction) {
+    throw new Error("登録しているアプリが見つかりません。");
+  }
+
+  if (appIntroduction.userId !== userId) {
+    throw new Error(
+      "登録しているユーザーのIDとフォームから送信されたユーザーのIDが一致しません。"
+    );
+  }
 
   try {
     await prisma.appIntroduction.delete({
       where: {
-        id,
+        id: appId,
       },
     });
     console.log("アプリの削除に成功しました。");
